requirements/base.txt
# Bases
psycopg2-binary==2.8.6

# Django

Django==3.1.3
django-allauth
django-celery-results
django-countries
behave-django
django-extensions
django-model-utils
django-filter==2.4.0
drf-yasg==1.20.0
djangorestframework
djangorestframework-jwt
django_select2
django-rest-auth
django-rest-swagger